Rename geometric_types.hpp to geometry.hpp
