+++ tests/test_configurationspace_construction.cpp
@@ -1,5 +1,5 @@
 #include <gtest/gtest.h>
-#include <BURST/geometric_types.hpp>
+#include <BURST/geometry.hpp>
 #include <BURST/configuration_space.hpp>
 #include <BURST/wall_space.hpp>
